+++ fastly/products/apidiscovery/doc.go
@@ -1,0 +1,3 @@
+// Package apidiscovery contains API operations to enable and
+// disable the API Discovery product on a service
+package apidiscovery